Primitive Fence Midsection Recipe Uses Very Thin Branches
Changed the primitive fence midsection recipe to use "branch-very-thin-*" as opposed to "branch-thin-*" to match the primitive fence support recipe.

(The primitive fence support recipe uses very thin branches to represent the thinner fence parts. As such, it is inconsistent that the midsection uses just thin branches when its fence parts are just as thin as the support's).