Fix Neolithic Backpack Problems
Block backpacks cause exception when loading a savegame, causing a loss of inventory. To combat this, vanilla backpacks (which are items and not blocks) have been re-enabled and the Neolithic block bags have been changed to drop the vanilla items instead of themselves. Also, backpack attribute has been removed from these block bags to prevent accidental use in the future.